Move TextureWhitePixel into its own file and make the class internal
